requirements.txt
ipywidgets==8.1.2
matplotlib==3.8.4
pandas==2.2.2
# torch==2.3.0
# torchvision==0.18.0
tqdm==4.66.4
# pymongo==4.10.1
neo4j==5.25.0
pyneo==0.2.0
py2neo==2021.2.4
apache-airflow-providers-postgres==6.3.0
kaggle
openpyxl>=3.1.0
streamlit
plotly